moved close btn&made it sketchier but prettier
