+++ src/Components/AppBar/Default/LoginForm.jsx
@@ -21,7 +21,10 @@ import {
 
 import LockOutlinedIcon from '@mui/icons-material/LockOutlined';
 import { Visibility, VisibilityOff } from '@mui/icons-material';
+import CloseIcon from '@mui/icons-material/Close';
+
 import AlertBox from '../../AlertBox';
+import Tooltip from '../../Tooltip';
 
 function LoginForm({ setCurrentOpenDrawer }) {
     const { register, handleSubmit } = useForm();
@@ -44,6 +47,43 @@ function LoginForm({ setCurrentOpenDrawer }) {
 
     const handleClickCloseDrawer = () => {
         setCurrentOpenDrawer('');
+    };
+
+    const CloseDrawerButton = () => {
+        function handleClick() {
+            setCurrentOpenDrawer('');
+        }
+
+        const buttonHover = {
+            '&:hover .MuiAvatar-root': {
+                backgroundColor: 'primary.dark',
+            },
+        };
+
+        return (
+            <Tooltip title="Sulje">
+                <IconButton
+                    onClick={handleClick}
+                    sx={[
+                        buttonHover,
+                        {
+                            p: '0',
+                            marginTop: '1rem',
+                        },
+                    ]}
+                >
+                    <Avatar
+                        sx={{
+                            bgcolor: 'primary.main',
+                            width: 48,
+                            height: 48,
+                        }}
+                    >
+                        <CloseIcon />
+                    </Avatar>
+                </IconButton>
+            </Tooltip>
+        );
     };
 
     return (
@@ -126,6 +166,7 @@ function LoginForm({ setCurrentOpenDrawer }) {
                     >
                         Luo uusi tunnus
                     </Button>
+                    <CloseDrawerButton />
                 </Box>
             </Container>
         </>
